+++ frontend/relative_meas_gen/launch/odomRelativeMeasSync.launch
@@ -1,10 +1,20 @@
 <launch>
     <!-- Setup arguments for node name -->
-    <arg name="odom_relative_meas_sync_node_name" default="odom_relative_meas_sync"/>
-
-    <!-- Setup parameters -->
-    <param name="odom_relative_meas_sync_node_name" value="$(arg odom_relative_meas_sync_node_name)"/>
+    <arg name="odom_relative_meas_sync_node_name"/>
+    <arg name="relative_meas_topic"/>
+    <arg name="odom_topic_observer"/>
+    <arg name="odom_topic_observed"/>
+    <arg name="observer_ID"/>
+    <arg name="observed_ID"/>
     
     <!-- Launch the node -->
-    <node pkg="relative_meas_gen" type="odomRelativeMeasSync.py" name="$(arg odom_relative_meas_sync_node_name)" output="screen"/>
+    <node pkg="relative_meas_gen" type="odomRelativeMeasSync.py" name="$(arg odom_relative_meas_sync_node_name)" output="screen">
+        <param name="odom_relative_meas_sync_node_name" value="$(arg odom_relative_meas_sync_node_name)"/>
+        <param name="relative_meas_topic" value="$(arg relative_meas_topic)"/>
+        <param name="odom_topic_observer" value="$(arg odom_topic_observer)"/>
+        <param name="odom_topic_observed" value="$(arg odom_topic_observed)"/>
+        <param name="observer_ID" value="$(arg observer_ID)"/>
+        <param name="observed_ID" value="$(arg observed_ID)"/>
+    </node>
+
 </launch>
